+++ sessions-core/src/test/java/com/adyen/checkout/sessions/core/internal/SessionComponentEventHandlerTest.kt
@@ -23,6 +23,7 @@ import com.adyen.checkout.core.internal.util.Logger
 import com.adyen.checkout.sessions.core.SessionComponentCallback
 import com.adyen.checkout.sessions.core.SessionModel
 import com.adyen.checkout.sessions.core.SessionPaymentResult
+import com.adyen.checkout.sessions.core.TestComponentState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow