+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 ## New
 - Add support for Multibanco voucher.
 - Permission request is now being delegated to the `ActionComponentCallback`, `SessionComponentCallback` or `ComponentCallback` to handle it and return result through callback.
+  - If not handled, a toast will be shown stating that permission is not granted.
 - For voucher actions which have no `url` or `downloadUrl`, "Save as image" option will be offered to save the Voucher in `Downloads` folder.
   - Vouchers will save an image to user's phone with the following name format "Payment method type" + "Formatted data and time" (e.g. multibanco-2024-01-09T16_41_10).
 - Creating configurations just became easier. Using a DSL you can now create configurations in a more declarative and concise way: